+++ src/render/ui/panel.h
@@ -31,8 +31,8 @@ namespace render::ui
 		void AddChild(const std::shared_ptr<Panel>& panel);
 		void AddModel(int x, int y, int width, int height, Mesh& model);
 
-		void SetWidth(float width);
-		void SetHeight(float height);
+		void SetWidth(int width);
+		void SetHeight(int height);
 
 		void ClearModels();
